+++ CHANGELOG.md
@@ -4,6 +4,15 @@
 available on:
  ▶️ https://www.npmjs.com/package/@chat21/chat21-http-server
 
+## v0.2.37
+- Added index { "app_id": 1, "conversWith": 1} to ChatDB to improve performance of deleteConversationsByConversWith()
+
+## v0.2.36
+- Added endpoint /:app_id/:group_id/conversations/timelines. This endpoint removes all the conversations, from all timelines - belonging to a specific group-id
+
+## v0.2.35
+- ?
+
 ## v0.2.34
 - Fixed sound: "default" for ios/android